another hold-over from int-to-size_t conversion
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2817 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/ten/tendSatin.c
@@ -111,7 +111,7 @@ tend_satinGen(Nrrd *nout, float parm, float mina, float maxa, int wsize,
   char me[]="tend_satinGen", err[AIR_STRLEN_MED], buff[AIR_STRLEN_SMALL];
   Nrrd *nconf, *neval, *nevec;
   float *conf, *eval, *evec;
-  int xi, yi, zi, size[3];
+  size_t xi, yi, zi, size[3];
   float x, y, z, min[3], max[3];
 
   if (torus) {
